@@ -1,6 +1,6 @@
 package com.apriori.bcs.utils;
 
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.fail;
 
 import com.apriori.bcs.controller.BatchPartResources;
 import com.apriori.bcs.models.response.Part;
