@@ -1,5 +1,8 @@
 package com.apriori;
 
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
 import com.apriori.dataservice.TestDataService;
 import com.apriori.pagedata.WorkFlowData;
 import com.apriori.pageobjects.login.CicLoginPage;
@@ -9,7 +12,6 @@ import com.apriori.reader.file.user.UserUtil;
 import com.apriori.testrail.TestRail;
 
 import io.qameta.allure.Description;
-import org.junit.Assert;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
@@ -40,8 +42,8 @@ public class WorkflowHistoryTests extends TestBaseUI {
             .selectScheduleTab()
             .selectWorkflowAndInvoke(workFlowData.getWorkflowName());
 
-        Assert.assertEquals("The job was successfully started!", workflowHome.getWorkFlowStatusMessage());
+        assertEquals("The job was successfully started!", workflowHome.getWorkFlowStatusMessage());
 
-        Assert.assertTrue("Verify Workflow job is finished",workflowHome.selectViewHistoryTab().searchAndTrackWorkFlowStatus(workFlowData.getWorkflowName()));
+        assertTrue(workflowHome.selectViewHistoryTab().searchAndTrackWorkFlowStatus(workFlowData.getWorkflowName()),"Verify Workflow job is finished");
     }
 }
